Created separate file for inital books state
