+++ src/redux/books/booksSlice.js
@@ -1,28 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-
-// Define the initial state for books
-export const initialBooksState = {
-  books: [
-    {
-      item_id: 'item1',
-      title: 'The Great Gatsby',
-      author: 'John Smith',
-      category: 'Fiction',
-    },
-    {
-      item_id: 'item2',
-      title: 'Anna Karenina',
-      author: 'Leo Tolstoy',
-      category: 'Fiction',
-    },
-    {
-      item_id: 'item3',
-      title: 'The Selfish Gene',
-      author: 'Richard Dawkins',
-      category: 'Nonfiction',
-    },
-  ],
-};
+import { initialBooksState } from './initialStateSlice';
 
 // Define the books slice
 const booksSlice = createSlice({
